requirements.txt
python-dotenv==1.0.1
loguru==0.7.2
translators==5.9.0
alembic==1.13.1
asyncpg==0.29.0
logtail-python==0.2.10
pytz==2024.1
APScheduler==3.10.4
mypy==1.9.0
PyMySQL==1.1.1
aiogram==3.12.0
beautifulsoup4>=4.10.0
selenium==4.23.1
webdriver-manager==4.0.2
